Generalize docset generation to aribtrary rustdocs
This is a huge pile of hacks, but it works.  The Rakefile now requires
some environment variables that set, in addition to the path to the
rustdoc output directory, the various names and identifiers used to name
and configure the docset.

Included is a script for generating the nightly docset and an example
script for would-be docset generators.  Eventually this capability
should be better-documented.

Close #3